requirements.txt
apache-beam[gcp]==2.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
requests==2.*